Add some auth and validations

+++ db/migrate/20221028103221_add_school_alternative_heating_sources.rb
@@ -3,7 +3,7 @@ class AddSchoolAlternativeHeatingSources < ActiveRecord::Migration[6.0]
     create_table :alternative_heating_sources do |t|
       t.references :school, foreign_key: true
       t.integer :source
-      t.numeric :percent_of_overall_use
+      t.integer :percent_of_overall_use
       t.text :notes
     end
   end
